furia_echo_backend/requirements.txt
Django>=5.2
djangorestframework
cloudscraper
fake_useragent
requests
python-dotenv
gunicorn
whitenoise
bs4